requirements.txt
Django
django-ckeditor
django-filter
django-suit
djangorestframework
inflect
Pillow
